fix module conflicts
