+++ lib/boardgame_engine/sample_games.rb
@@ -1,0 +1,5 @@
+require_relative "chess"
+require_relative "connect4"
+
+module SampleGames
+end
